pyproject.toml
[project]
name = "paper-analysis-agent"
version = "0.1.0"
description = "Academic Paper Analysis Agent using LangGraph"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "arxiv",
    "pypdf",
    "python-dotenv",
    "tiktoken",
    "streamlit",
    "langchain-anthropic>=0.3.22",
    "pymupdf>=1.26.5",
    "tavily-python>=0.7.21",
    "langchain-exa>=0.3.1",
    "langchain-tavily>=0.2.11",
    "langchain-text-splitters>=0.3.11",
    "google-search-results>=2.4.2",
    "pymupdf4llm>=0.0.27",
    "langfuse>=3.7.0",
]

[tool.uv]
package = true
